Fix compilation issues on Sonic

+++ app/interior_tet_opt/src/interior_tet_opt/Mesh.cpp
@@ -419,7 +419,7 @@ void InteriorTetOpt::collapse_all_edges(bool is_limit_length)
             executor(*this, collect_all_ops);
             wmtk::logger().info(
                 "Collapsed {}, retrying failed {}",
-                count_success,
+                (int)count_success,
                 collect_failure_ops.size());
             collect_all_ops.clear();
             for (auto& item : collect_failure_ops) collect_all_ops.emplace_back(item);
